requirements.txt
# =============================================================================
# CAPTAIN RACCOON — DEPENDENCIES
# =============================================================================

# ── Core AI / LLM ─────────────────────────────────────────────────────────────
anthropic>=0.40.0
replicate>=0.32.0
huggingface-hub>=0.27.0

# ── Azure (TTS + Storage) ─────────────────────────────────────────────────────
azure-cognitiveservices-speech>=1.40.0
azure-storage-blob>=12.24.0

# ── Google Cloud (TTS fallback) ───────────────────────────────────────────────
google-cloud-texttospeech>=2.17.0

# ── YouTube ───────────────────────────────────────────────────────────────────
google-api-python-client>=2.160.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ── MongoDB ───────────────────────────────────────────────────────────────────
pymongo>=4.10.0
motor>=3.6.0

# ── AWS / DigitalOcean Spaces (S3-compatible) ──────────────────────────────────
boto3>=1.35.0

# ── HTTP ──────────────────────────────────────────────────────────────────────
httpx>=0.28.0
requests>=2.32.0

# ── Config + Validation ───────────────────────────────────────────────────────
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0.2
python-dotenv>=1.0.1

# ── Retry ─────────────────────────────────────────────────────────────────────
tenacity>=9.0.0

# ── Monitoring ────────────────────────────────────────────────────────────────
sentry-sdk[httpx]>=2.19.0
datadog>=0.50.0
newrelic>=10.3.0

# ── Structured Logging ────────────────────────────────────────────────────────
structlog>=24.4.0

# ── Scheduling ────────────────────────────────────────────────────────────────
schedule>=1.2.2

# ── Audio ─────────────────────────────────────────────────────────────────────
pydub>=0.25.1

# ── Image Utils ───────────────────────────────────────────────────────────────
pillow>=11.0.0

# ── CLI ───────────────────────────────────────────────────────────────────────
rich>=13.9.0
click>=8.1.8

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-mock>=3.14.0
